Cleaning css formats and class names for robotics, and fixed the header to look a little prettier.
I chose a bold color for the header, but ill probably change it to a
gray-ish color. Actually is two colors.

The css class names are snake cased.

The css format has changed to have every attribute on its own line to
match my brilliant partner's work.

The logo had a lot of touch-ups. Again, Ive spent entirely too much time
on this dumb logo.

Going to keep working. This is a "half-way" commit to prevent loss of
work.