"Who we are" responsive fix. Clear code added to css defaults.
Okay so.....sometimes, when this issue happens, a random div has to be
placed after the problem area with the css attribute "clear:both;"

Added a class called "clear" that will allow for a quick fix for this in
the future.

PS - Use this class when a parent wont properly hold the height of its
child block element.